Move agent Delegated mode toggle to the Advanced tab
