+++ frontend/apps/console/src/features/agents/components/edit-agent/advanced-settings/EditAdvancedSettings.tsx
@@ -16,14 +16,17 @@
  * under the License.
  */
 
-import {Stack} from '@wso2/oxygen-ui';
+import {FormControlLabel, Stack, Switch} from '@wso2/oxygen-ui';
+import {useTranslation} from 'react-i18next';
 import AllowedUserTypesSection from './AllowedUserTypesSection';
 import OperationModesSection from './OperationModesSection';
 import OwnerSection from './OwnerSection';
 import SecuritySection from './SecuritySection';
 import TokenEndpointAuthMethodSection from './TokenEndpointAuthMethodSection';
 import AudienceSection from '../../../../applications/components/edit-application/advanced-settings/AudienceSection';
-import type {OAuth2Config} from '../../../../applications/models/oauth';
+import {OAuth2GrantTypes, type OAuth2Config} from '../../../../applications/models/oauth';
+import {applyGrantTypesChange} from '../../../../applications/utils/oauth2Rules';
+import {DELEGATED_ONLY_GRANTS} from '../../../constants/delegationGrants';
 import type {Agent, AgentInboundAuthConfig, OAuthAgentConfig} from '../../../models/agent';
 
 interface EditAdvancedSettingsProps {
@@ -39,12 +42,31 @@ export default function EditAdvancedSettings({
   oauth2Config = undefined,
   onFieldChange,
 }: EditAdvancedSettingsProps) {
+  const {t} = useTranslation();
+  const isUnlocked = oauth2Config?.grantTypes?.includes(OAuth2GrantTypes.AUTHORIZATION_CODE) ?? false;
+
   const handleOAuth2ConfigChange = (updates: Partial<OAuth2Config>) => {
     const currentInboundAuth: AgentInboundAuthConfig[] = editedAgent.inboundAuthConfig ?? agent.inboundAuthConfig ?? [];
     const updatedInboundAuth = currentInboundAuth.map((auth) =>
       auth.type === 'oauth2' ? {...auth, config: {...auth.config, ...updates} as OAuthAgentConfig} : auth,
     );
     onFieldChange('inboundAuthConfig', updatedInboundAuth);
+  };
+
+  // Delegated mode unlocks the delegated-only grants below and the Flows/Tokens tabs. Toggling it
+  // just flips authorization_code on/off; applyGrantTypesChange handles the dependent grants.
+  const handleDelegationToggle = (checked: boolean): void => {
+    if (!oauth2Config || checked === isUnlocked) return;
+    const grantTypes = oauth2Config.grantTypes ?? [];
+    const nextGrantTypes = checked
+      ? [...new Set([...grantTypes, OAuth2GrantTypes.AUTHORIZATION_CODE])]
+      : grantTypes.filter((grant) => !DELEGATED_ONLY_GRANTS.includes(grant));
+    const updates = applyGrantTypesChange(oauth2Config, nextGrantTypes);
+    // PKCE is fully derived from authorization_code for agents.
+    if (checked) {
+      updates.pkceRequired = true;
+    }
+    handleOAuth2ConfigChange(updates);
   };
 
   const handleDefaultAudienceChange = (audience: string) => {
@@ -58,6 +80,16 @@ export default function EditAdvancedSettings({
 
   return (
     <Stack spacing={3}>
+      <FormControlLabel
+        control={
+          <Switch
+            checked={isUnlocked}
+            onChange={(e) => handleDelegationToggle(e.target.checked)}
+            disabled={!oauth2Config || agent.isReadOnly === true}
+          />
+        }
+        label={t('agents:edit.advanced.delegationToggle.label', 'Delegated mode')}
+      />
       <OwnerSection agent={agent} editedAgent={editedAgent} onFieldChange={onFieldChange} />
       <AllowedUserTypesSection
         agent={agent}
