+++ frontend/apps/console/src/features/agents/components/edit-agent/flows/__tests__/EditFlowsSettings.test.tsx
@@ -17,7 +17,6 @@
  */
 
 import {render, screen} from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
 import {describe, it, expect, vi} from 'vitest';
 import type {Application} from '../../../../../applications/models/application';
 import type {Agent} from '../../../../models/agent';
@@ -99,108 +98,5 @@ describe('EditFlowsSettings', () => {
     );
 
     expect(screen.getByTestId('auth-flow')).toHaveAttribute('data-readonly', 'true');
-  });
-
-  describe('Delegated mode toggle', () => {
-    it('shows the toggle checked when Delegated mode is on', () => {
-      render(
-        <EditFlowsSettings
-          agent={baseAgent}
-          editedAgent={{}}
-          oauth2Config={{grantTypes: ['authorization_code'], responseTypes: ['code']}}
-          onFieldChange={mockOnFieldChange}
-        />,
-      );
-
-      expect(screen.getByRole('switch', {name: 'Delegated mode'})).toBeChecked();
-    });
-
-    it('shows the toggle unchecked when Delegated mode is off', () => {
-      render(
-        <EditFlowsSettings
-          agent={baseAgent}
-          editedAgent={{}}
-          oauth2Config={{grantTypes: ['client_credentials'], responseTypes: []}}
-          onFieldChange={mockOnFieldChange}
-        />,
-      );
-
-      expect(screen.getByRole('switch', {name: 'Delegated mode'})).not.toBeChecked();
-    });
-
-    it('turns on Delegated mode by adding authorization_code and requiring PKCE', async () => {
-      const user = userEvent.setup();
-      render(
-        <EditFlowsSettings
-          agent={{
-            ...baseAgent,
-            inboundAuthConfig: [{type: 'oauth2', config: {grantTypes: ['client_credentials'], responseTypes: []}}],
-          }}
-          editedAgent={{}}
-          oauth2Config={{grantTypes: ['client_credentials'], responseTypes: []}}
-          onFieldChange={mockOnFieldChange}
-        />,
-      );
-
-      await user.click(screen.getByRole('switch', {name: 'Delegated mode'}));
-
-      expect(mockOnFieldChange).toHaveBeenCalledWith(
-        'inboundAuthConfig',
-        expect.arrayContaining([
-          expect.objectContaining({
-            type: 'oauth2',
-            config: expect.objectContaining({
-              grantTypes: expect.arrayContaining(['client_credentials', 'authorization_code']) as string[],
-              pkceRequired: true,
-            }) as Record<string, unknown>,
-          }),
-        ]),
-      );
-    });
-
-    it('turns off Delegated mode by dropping the delegated-only grants', async () => {
-      const user = userEvent.setup();
-      render(
-        <EditFlowsSettings
-          agent={{
-            ...baseAgent,
-            inboundAuthConfig: [
-              {
-                type: 'oauth2',
-                config: {grantTypes: ['client_credentials', 'authorization_code'], responseTypes: ['code']},
-              },
-            ],
-          }}
-          editedAgent={{}}
-          oauth2Config={{grantTypes: ['client_credentials', 'authorization_code'], responseTypes: ['code']}}
-          onFieldChange={mockOnFieldChange}
-        />,
-      );
-
-      await user.click(screen.getByRole('switch', {name: 'Delegated mode'}));
-
-      expect(mockOnFieldChange).toHaveBeenCalledWith(
-        'inboundAuthConfig',
-        expect.arrayContaining([
-          expect.objectContaining({
-            type: 'oauth2',
-            config: expect.objectContaining({grantTypes: ['client_credentials']}) as Record<string, unknown>,
-          }),
-        ]),
-      );
-    });
-
-    it('disables the toggle for read-only agents', () => {
-      render(
-        <EditFlowsSettings
-          agent={{...baseAgent, isReadOnly: true}}
-          editedAgent={{}}
-          oauth2Config={{grantTypes: ['authorization_code'], responseTypes: ['code']}}
-          onFieldChange={mockOnFieldChange}
-        />,
-      );
-
-      expect(screen.getByRole('switch', {name: 'Delegated mode'})).toBeDisabled();
-    });
   });
 });
